requirements.txt
ultralytics
opencv-python-headless
Flask 
opencv-python 
pillow 
numpy
azure-storage-blob
openai
openpyxl
python-dotenv
google-cloud-vision
google-api-python-client
requests
scikit-learn


